Initial 0.6.0 branch with traits
